use 24h format for releaseDirname instead of 12h
that is what was used previously, when 'moment' was used for date
formatting. 12h format causes terrible consequences:

(1) every timestamp repeats twice a day, for AM and PM. Which is bad
because deploy:cleanup task makes sorted list of directories, and
removes oldest, which might actually happen to be the directory which
was just released.

(2) if your project haven't specified exact version of grunt-shipit in
package.json, and multiple people are deploying from different
workspaces, one might be using 0.4.2 (12h) the other 0.4.0 (24h) and
then the cleanup of your current dir will hit you for sure.