Remove default reference to port 22.
Default reference to port 22 when one is not specified breaks default configuration in .ssh/config. This change removes the default refernce and uses the '-p' switch only when the port is explicitly specified. Otherwise the ssh command will fallback to default settings in .ssh/config or to port 22 when there is no local configuration for given hostname.